refactor: centralize answer formatting
